fix(agent): fail-closed on a stale-run wake-up dispatch failure + hardening
The reaper reclaims a stale RUNNING run onto the queue (RUNNING -> QUEUED)
and then dispatches an AgentRunQueuedMessage to wake a worker. On an async
transport a failed dispatch left the row QUEUED forever: the reaper only
re-finds stale RUNNING rows, never QUEUED, and no message would ever arrive,
so nothing would pick it up. Mirror AgentRuntime::enqueue()'s fail-closed
seam and settle the run FAILED instead of stranding an orphan.

Release-readiness hardening bundled in:

- AgentRuntime::dispatchRequeue: 2 ** $n widens to int|float; cast the
  capped backoff to int for the DelayStamp (exact, the cap bounds it).
- FalStorageGate: ResourceStorage::getFile() is nullable; guard null
  before checkFileActionPermission() rather than relying on a suppression.

Test coverage for previously unexercised branches:

- ReapStaleAgentRunsCommand: the dispatch-failure -> settle-FAILED path.
- ContextWindowManager: the non-positive-budget passthrough (reserve larger
  than the whole window) defers the transcript to the provider untouched.
- WaitingRunViewFactory: the NUMBER control arm (step "any" / decimal),
  enum -> string-coerced options, and the humanised label fallback.

+++ Classes/Service/Agent/AgentRuntime.php
@@ -312,7 +312,9 @@ final readonly class AgentRuntime implements AgentRuntimeInterface
             throw RunEnqueueFailedException::forRun($uuid);
         }
 
-        $delayMs = min(self::REQUEUE_BACKOFF_MS * (2 ** $priorRequeueCount), self::REQUEUE_BACKOFF_CAP_MS);
+        // 2 ** $n widens to int|float under static analysis; the cap keeps the
+        // result well inside int range, so the cast is exact, not lossy.
+        $delayMs = (int)min(self::REQUEUE_BACKOFF_MS * (2 ** $priorRequeueCount), self::REQUEUE_BACKOFF_CAP_MS);
         $this->messageBus->dispatch(new AgentRunQueuedMessage($uuid), [new DelayStamp($delayMs)]);
     }
